moved rule set resource to permissions
