+++ permissions/resource_rule_set.go
@@ -1,4 +1,4 @@
-package rule_sets
+package permissions
 
 import (
 	"context"
